rf: move erf and pro_density into its own files
